Changed Pixel A1R5G5B5 data type from Short to Int
Since VS demands a casting, it's easier for everyone just use int

+++ Bitmaps/A1R5G5B5Image.cs
@@ -29,7 +29,7 @@ namespace GaloreWare.Drawing.Bitmaps
         {
             Pixel tmp = new Pixel(0, 0, 0, 255);
             
-            short data;
+            int data;
             
             int i=0;
